pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "urlscraper"
version = "0.0.1"
description = "Workbench module to download one row per URL"
authors = ["Adam Hooper <adam@adamhooper.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
cjwmodule = "~=3.1"
cjwparquet = "~=2.1.0"
pandas = "~=0.25.3"
pyarrow = "~=2.0"
python = "~=3.8"
rfc3987 = "~=1.3"

[tool.poetry.dev-dependencies]
pytest = "~=6.0"
pytest-cov = "~=2.11"
pytest-freezegun = "~=0.4.2"
pytest-httpx = "~=0.10.1"

[tool.isort]
# Black compatibility
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
skipsdist = True
envlist = py38-{pyflakes,black,isort,pytest}

[flake8]
exclude = venv/*,tox/*,specs/*,build/*
ignore = E123,E128,E266,E402,W503,E731,W601
max-line-length = 88

[testenv]
skip_install = true
deps =
  pyflakes: pyflakes>=2.2
  black: black
  isort: isort
  pytest: poetry
commands_pre =
  pytest: poetry install -v
commands =
  pyflakes: pyflakes urlscraper.py tests
  black: black --check urlscraper.py tests
  isort: isort --check --diff urlscraper.py tests
  pytest: poetry run pytest --cov=urlscraper --cov-report term-missing -v
"""
